+++ src/odin/enhancedtrippath.cc
@@ -164,6 +164,14 @@ const std::string& TripLeg_Sidewalk_Name(int v) {
   return f->second;
 }
 
+bool is_forward(uint32_t turn_degree) {
+  return ((turn_degree > 314) || (turn_degree < 46));
+}
+
+bool is_wider_forward(uint32_t turn_degree) {
+  return ((turn_degree > 304) || (turn_degree < 56));
+}
+
 } // namespace
 
 namespace valhalla {
@@ -409,11 +417,11 @@ bool EnhancedTripLeg_Edge::IsOneway() const {
 }
 
 bool EnhancedTripLeg_Edge::IsForward(uint32_t prev2curr_turn_degree) const {
-  return ((prev2curr_turn_degree > 314) || (prev2curr_turn_degree < 46));
+  return is_forward(prev2curr_turn_degree);
 }
 
 bool EnhancedTripLeg_Edge::IsWiderForward(uint32_t prev2curr_turn_degree) const {
-  return ((prev2curr_turn_degree > 304) || (prev2curr_turn_degree < 56));
+  return is_wider_forward(prev2curr_turn_degree);
 }
 
 bool EnhancedTripLeg_Edge::IsStraightest(uint32_t prev2curr_turn_degree,
@@ -1195,6 +1203,10 @@ bool EnhancedTripLeg_IntersectingEdge::IsTraversableOutbound(
   return false;
 }
 
+bool EnhancedTripLeg_IntersectingEdge::IsHighway() const {
+  return ((road_class() == TripLeg_RoadClass_kMotorway) && !(use() == TripLeg_Use_kRampUse));
+}
+
 std::string EnhancedTripLeg_IntersectingEdge::ToString() const {
   std::string str;
   str.reserve(128);
@@ -1331,9 +1343,7 @@ void EnhancedTripLeg_Node::CalculateRightLeftIntersectingEdgeCounts(
 bool EnhancedTripLeg_Node::HasFowardIntersectingEdge(uint32_t from_heading) {
 
   for (int i = 0; i < intersecting_edge_size(); ++i) {
-    uint32_t intersecting_turn_degree =
-        GetTurnDegree(from_heading, intersecting_edge(i).begin_heading());
-    if ((intersecting_turn_degree > 314) || (intersecting_turn_degree < 46)) {
+    if (is_forward(GetTurnDegree(from_heading, intersecting_edge(i).begin_heading()))) {
       return true;
     }
   }
@@ -1345,11 +1355,35 @@ bool EnhancedTripLeg_Node::HasForwardTraversableIntersectingEdge(
     const TripLeg_TravelMode travel_mode) {
 
   for (int i = 0; i < intersecting_edge_size(); ++i) {
-    uint32_t intersecting_turn_degree =
-        GetTurnDegree(from_heading, intersecting_edge(i).begin_heading());
-    bool xedge_traversable_outbound = GetIntersectingEdge(i)->IsTraversableOutbound(travel_mode);
-    if (((intersecting_turn_degree > 314) || (intersecting_turn_degree < 46)) &&
-        xedge_traversable_outbound) {
+    if (is_forward(GetTurnDegree(from_heading, intersecting_edge(i).begin_heading())) &&
+        GetIntersectingEdge(i)->IsTraversableOutbound(travel_mode)) {
+      return true;
+    }
+  }
+  return false;
+}
+
+bool EnhancedTripLeg_Node::HasWiderForwardTraversableIntersectingEdge(
+    uint32_t from_heading,
+    const TripLeg_TravelMode travel_mode) {
+
+  for (int i = 0; i < intersecting_edge_size(); ++i) {
+    if (is_wider_forward(GetTurnDegree(from_heading, intersecting_edge(i).begin_heading())) &&
+        GetIntersectingEdge(i)->IsTraversableOutbound(travel_mode)) {
+      return true;
+    }
+  }
+  return false;
+}
+
+bool EnhancedTripLeg_Node::HasWiderForwardTraversableHighwayXEdge(
+    uint32_t from_heading,
+    const TripLeg_TravelMode travel_mode) {
+
+  for (int i = 0; i < intersecting_edge_size(); ++i) {
+    auto xedge = GetIntersectingEdge(i);
+    if (is_wider_forward(GetTurnDegree(from_heading, xedge->begin_heading())) &&
+        xedge->IsTraversableOutbound(travel_mode) && xedge->IsHighway()) {
       return true;
     }
   }