+++ valhalla/odin/enhancedtrippath.h
@@ -469,6 +469,8 @@ public:
 
   bool IsTraversableOutbound(const TripLeg_TravelMode travel_mode) const;
 
+  bool IsHighway() const;
+
   std::string ToString() const;
 
 protected:
@@ -584,6 +586,12 @@ public:
 
   bool HasForwardTraversableIntersectingEdge(uint32_t from_heading,
                                              const TripLeg_TravelMode travel_mode);
+
+  bool HasWiderForwardTraversableIntersectingEdge(uint32_t from_heading,
+                                                  const TripLeg_TravelMode travel_mode);
+
+  bool HasWiderForwardTraversableHighwayXEdge(uint32_t from_heading,
+                                              const TripLeg_TravelMode travel_mode);
 
   bool HasTraversableOutboundIntersectingEdge(const TripLeg_TravelMode travel_mode);